Roll-up v4.1 changes  to master

+++ fog/ingest/server/tests/tx_recovery.rs
@@ -111,7 +111,7 @@ fn test_ingest_polling_integration(base_port: u16, mut rng: RngType, logger: Log
 // Function for generating the random number of txs to put in a block
 fn gen_num_tx_for_block(rng: &mut impl RngCore) -> usize {
     loop {
-        let n = (rng.next_u64() % NUM_TX_PER_BLOCK as u64) as usize;
+        let n = (rng.next_u64() % NUM_TX_PER_BLOCK) as usize;
         if n != 0 {
             break n;
         }
